Removed unnecessary folder
