--- recipes/oatpp/1.1.0/CMakeLists.txt
@@ -1,7 +1,0 @@
-cmake_minimum_required(VERSION 2.8.12)
-project(cmake_wrapper)
-
-include("conanbuildinfo.cmake")
-conan_basic_setup()
-
-add_subdirectory(source_subfolder)
