--- recipes/oatpp/1.1.0/conandata.yml
@@ -1,4 +1,0 @@
-sources:
-  "1.1.0":
-    url: "https://github.com/oatpp/oatpp/archive/1.1.0.tar.gz"
-    sha256: "fcc59615e1282d51826093f95b6c0ded3240ea5c1f322bc253bc36da1f0d3c1c"
