--- recipes/oatpp/1.1.0/test_package/conanfile.py
@@ -1,18 +1,0 @@
-import os
-
-from conans import ConanFile, CMake, tools
-
-
-class OatppTestConan(ConanFile):
-    settings = "os", "compiler", "build_type", "arch"
-    generators = "cmake"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
-
-    def test(self):
-        if not tools.cross_building(self):
-            bin_path = os.path.join("bin", "test_package")
-            self.run(bin_path, run_environment=True)
