+++ recipes/oatpp/config.yml
@@ -2,4 +2,4 @@ versions:
   "1.0.0":
     folder: all
   "1.1.0":
-    folder: 1.1.0
+    folder: all
